testfn/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "testfn"
version = "0.0.1"
description = "Comprehensive self-hosted testing platform for developers (Python SDK)"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["testing", "test-runner", "automation", "superfunctions"]
authors = [
  { name = "21n" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Testing",
]
dependencies = [
  "pydantic>=2.5.0",
  "sqlalchemy>=2.0.0",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "mypy>=1.7.0",
  "ruff>=0.1.6",
  "black>=23.11.0",
]

[project.urls]
Documentation = "https://docs.superfunctions.dev/testfn"
Repository = "https://github.com/21nCo/super-functions"
Issues = "https://github.com/21nCo/super-functions/issues"

[tool.hatch.build.targets.wheel]
packages = ["testfn"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
]
ignore = [
  "E501", # line too long, handled by black
]
